Fix wrapping in desktop mode
The wrapping of the Yes/No/? column header wasn't matching the numbers
for the actual events. It turned out that this margin was being applied
to the cell in the event row but not in the header, so the event row was
wider so it wrapped earlier than the column header. Removing it doesn't
seem to make any difference, so doing that.