Fix js id template for new uri escaping
In 269c543243 I fixed the uri escaping to just automatically
percent-encode all the parameters (better, safer). But the event link
"id={{id}}" goes to the javascript in goc2.js to be filled in with the
actual event, and the javascript wasn't looking for %7B%&bid%7D%7D. I
feel like there's a more correct way to handle this, but it's not coming
to mind at the moment, and this is good enough for now.