Add some indexes on the event table
After worrying about the performance of that new prev/next query with the
windowing functions, I realized the event table  could use some indexes.  (note
that pragma index_list("event") does *not* show the index on the PRIMARY KEY
column, see https://stackoverflow.com/a/55036663/514032)

This improves the query plan somewhat:

BEFORE:

    QUERY PLAN
    |--CO-ROUTINE SUBQUERY 1
    |  |--CO-ROUTINE SUBQUERY 3
    |  |  |--CO-ROUTINE SUBQUERY 4
    |  |  |  |--SCAN event
    |  |  |  `--USE TEMP B-TREE FOR ORDER BY
    |  |  `--SCAN SUBQUERY 4
    |  |--SCAN SUBQUERY 3
    |  `--USE TEMP B-TREE FOR ORDER BY
    `--SCAN SUBQUERY 1

AFTER:

    QUERY PLAN
    |--CO-ROUTINE SUBQUERY 1
    |  |--CO-ROUTINE SUBQUERY 3
    |  |  |--CO-ROUTINE SUBQUERY 4
    |  |  |  |--SEARCH event USING INDEX event_type_idx (type=?)
    |  |  |  `--USE TEMP B-TREE FOR ORDER BY
    |  |  `--SCAN SUBQUERY 4
    |  |--SCAN SUBQUERY 3
    |  `--USE TEMP B-TREE FOR ORDER BY
    `--SCAN SUBQUERY 1

but really there's still only a hundred events in the table. I suppose the
ultimate solution would be to store a linked list and update the links whenever
an event is created or its date is changed.